service/requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-cloud-bigquery==3.13.0
google-cloud-secret-manager==2.16.0
requests==2.31.0
tenacity==8.2.3
python-dotenv==1.0.0
